--- packages/desktop_security_center_grpc/go.mod
@@ -1,12 +1,0 @@
-module github.com/canonical/desktop-security-center/packages/desktop_security_center_grpc
-
-go 1.21
-
-require (
-	github.com/coreos/go-systemd v0.0.0-20191104093116-d3cd4ed1dbcf // indirect
-	github.com/golang/protobuf v1.5.3 // indirect
-	golang.org/x/sys v0.13.0 // indirect
-	google.golang.org/genproto/googleapis/rpc v0.0.0-20231002182017-d307bd883b97 // indirect
-	google.golang.org/grpc v1.60.1 // indirect
-	google.golang.org/protobuf v1.31.0 // indirect
-)
