+++ packages/systemd/server.service
@@ -1,0 +1,12 @@
+[Unit]
+Description=Whatever
+Requires=network.target
+After=multi-user.target
+
+[Service]
+Type=simple
+ExecStart=/usr/local/bin/seccenter
+
+[Install]
+WantedBy=multi-user.target
+
